Add incident reporting feature and refactor routes
Updated AppShell to change login route to lowercase and added a new route for reporting incidents. Registered the ReportIncidentViewModel and ReportIncidentView in MauiProgram. Implemented CreateIncidentAsync in IncidentService for incident creation with authentication checks. Modified HomeMapViewModel for navigation to the new report incident route and updated toolbar items in HomeMapView. Added CreateIncidentRequest class and created ReportIncidentView with its ViewModel to manage incident reporting functionality.